+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to Semantic Versioning.
 ### Added
 
 * Stencil buffers can now be used, both with the main framebuffer and with `Canvas` rendering. ([@tesselode](https://github.com/tesselode) in [#261](https://github.com/17cupsofcoffee/tetra/pull/261))
+    * Stencil buffers will only be created if they are requested via `ContextBuilder::stencil_buffer` or `CanvasBuilder::stencil_buffer`.
 * `window::get_physical_size` has been added, making it possible to find out the size of the window in unscaled, physical pixels.
 * `window::get_dpi_scale` has been added, allowing the window/monitor's pixel density to be queried. ([@sumibi-yakitori](https://github.com/sumibi-yakitori) in [#249](https://github.com/17cupsofcoffee/tetra/pull/249))
 * `Texture` and `Canvas` now have a `get_data` method, allowing their texture data to be copied from the GPU into an `ImageData`.
@@ -36,7 +37,7 @@ This project adheres to Semantic Versioning.
 ### Deprecated
 
 * `ImageData::from_rgba` and `BmFontBuilder::with_page_rgba` has been deprecated in favour of more consistently named functions (see above).
-* `Canvas::multisampled` has been deprecated in favour of `Canvas::with_settings`.
+* `Canvas::multisampled` has been deprecated in favour of `Canvas::builder`.
 
 ## [0.6.3] - 2021-04-09